requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diskcache==5.6.3
fastapi==0.109.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.3
llama_cpp_python==0.2.30
MarkupSafe==2.1.3
numpy==1.26.3
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
